+++ autokeras/image/image_supervised.py
@@ -3,10 +3,12 @@ from abc import ABC
 import numpy as np
 from multiprocessing import Pool, cpu_count
 
+from autokeras.backend import Backend
 from autokeras.constant import Constant
 from autokeras.nn.loss_function import classification_loss, regression_loss
 from autokeras.nn.metric import Accuracy, MSE
-from autokeras.preprocessor import OneHotEncoder, ImageDataTransformer
+from autokeras.preprocessor import OneHotEncoder
+from autokeras.backend.torch import ImageDataTransformer
 from autokeras.supervised import PortableDeepSupervised, DeepTaskSupervised
 from autokeras.utils import pickle_to_file, \
     read_csv_file, read_image, compute_image_resize_params, resize_image_data
@@ -161,8 +163,7 @@ class ImageSupervised(DeepTaskSupervised, ABC):
             x: DATA TO TRANSFORM
         """
         if self.data_transformer is None:
-            self.data_transformer = ImageDataTransformer(
-                x, augment=self.augment)
+            self.data_transformer = Backend.get_image_transformer(x, augment=self.augment)
 
     def preprocess(self, x):
         """THIS FUNCTION NEEDS A DESCRIPTION.
