+++ examples/net_modules/mlp_module.py
@@ -4,7 +4,8 @@ from keras.datasets import mnist
 from autokeras import MlpModule
 from autokeras.nn.loss_function import classification_loss
 from autokeras.nn.metric import Accuracy
-from autokeras.preprocessor import OneHotEncoder, DataTransformerMlp
+from autokeras.preprocessor import OneHotEncoder
+from autokeras.backend.torch import DataTransformerMlp
 
 
 def transform_y(y_train):
